Oops, missed this.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@4651 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ protocolgame.cpp
@@ -524,7 +524,7 @@ void ProtocolGame::parsePacket(NetworkMessage &msg)
 
 	if(msg.size() == NETWORK_MAX_SIZE)
 	{
-		if(++m_maxSizeCount >= g_config.getNumber(ConfigManager::ALLOWED_MAX_PACKETS))
+		if(++m_maxSizeCount >= (uint32_t)g_config.getNumber(ConfigManager::ALLOWED_MAX_PACKETS))
 		{
 			IOBan::getInstance()->addIpBanishment(player->getIP(), -1, 0, "Massive packet sending", 0);
 			return;
